cleaned up injection logic for auth and billman
