+++ auth/src/main/kotlin/com/weesnerdevelopment/auth/database/AuthDatabase.kt
@@ -1,0 +1,21 @@
+package com.weesnerdevelopment.auth.database
+
+import auth.UsersTable
+import com.weesnerdevelopment.auth.user.UserHistoryTable
+import com.weesnerdevelopment.businessRules.WeesnerDevelopmentDatabase
+import history.HistoryTable
+import org.jetbrains.exposed.sql.SchemaUtils.create
+import org.jetbrains.exposed.sql.transactions.transaction
+
+interface AuthDatabase : WeesnerDevelopmentDatabase {
+    fun createTables(addLogger: Boolean) {
+        transaction {
+            create(
+                UsersTable,
+                UserHistoryTable,
+
+                HistoryTable
+            )
+        }
+    }
+}
