+++ billMan/src/main/kotlin/com/weesnerdevelopment/billman/database/BillManDatabase.kt
@@ -1,4 +1,4 @@
-package com.weesnerdevelopment.billman
+package com.weesnerdevelopment.billman.database
 
 import com.weesnerdevelopment.billman.bill.BillHistoryTable
 import com.weesnerdevelopment.billman.bill.BillSharedUsersTable
@@ -18,8 +18,7 @@ import com.weesnerdevelopment.billman.income.IncomeHistoryTable
 import com.weesnerdevelopment.billman.income.IncomeTable
 import com.weesnerdevelopment.billman.income.occurrence.IncomeOccurrenceHistoryTable
 import com.weesnerdevelopment.billman.income.occurrence.IncomeOccurrenceTable
-import com.zaxxer.hikari.HikariConfig
-import com.zaxxer.hikari.HikariDataSource
+import com.weesnerdevelopment.businessRules.WeesnerDevelopmentDatabase
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.SchemaUtils.create
 import org.jetbrains.exposed.sql.StdOutSqlLogger
@@ -27,18 +26,13 @@ import org.jetbrains.exposed.sql.addLogger
 import org.jetbrains.exposed.sql.transactions.transaction
 
 /**
- * The database factory for weesnerDevelopment.com.
+ * The database factory for weesnerDevelopment.com/billMan.
  */
-object BillManDatabase {
+interface BillManDatabase : WeesnerDevelopmentDatabase {
     /**
-     * Initializes the [Database] and creates the database tables if needed.
+     * Creates the [Database] tables.
      */
-    fun init(testing: Boolean) {
-        if (testing)
-            Database.connect("jdbc:h2:mem:test;DB_CLOSE_DELAY=-1", driver = "org.h2.Driver")
-        else
-            Database.connect(hikari())
-
+    fun createTables(addLogger: Boolean) {
         transaction {
             create(
                 BillTable,
@@ -67,19 +61,8 @@ object BillManDatabase {
                 IncomeOccurrenceHistoryTable
             )
 
-            if (testing)
+            if (addLogger)
                 addLogger(StdOutSqlLogger)
         }
     }
-
-    private fun hikari() = HikariDataSource(
-        HikariConfig().apply {
-            driverClassName = "org.h2.Driver"
-            jdbcUrl = "jdbc:h2:./server/billMan-database;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false"
-            maximumPoolSize = 3
-            isAutoCommit = false
-            transactionIsolation = "TRANSACTION_REPEATABLE_READ"
-            validate()
-        }
-    )
 }
