+++ businessRules/src/main/kotlin/com/weesnerdevelopment/businessRules/Server.kt
@@ -1,0 +1,7 @@
+package com.weesnerdevelopment.businessRules
+
+import io.ktor.application.*
+
+interface Server {
+    fun start(app: Application)
+}
