+++ businessRules/src/main/kotlin/com/weesnerdevelopment/businessRules/WeesnerDevelopmentDatabase.kt
@@ -1,0 +1,8 @@
+package com.weesnerdevelopment.businessRules
+
+interface WeesnerDevelopmentDatabase {
+    /**
+     * Initializes the [Database].
+     */
+    fun setup()
+}
